backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
groq
python-dotenv
supabase
pydantic
python-multipart
